feature : 화폐 타입을 따로 지정할 경우 : totalBudget, paymentDtoList 의 화폐 타입과 금액을 변환

+++ src/main/java/com/keypoint/keypointtravel/campaign/entity/CampaignBudget.java
@@ -30,7 +30,7 @@ public class CampaignBudget {
     private String category;
 
     @Column(nullable = false)
-    private Long amount;
+    private float amount;
 
     @Builder
     public CampaignBudget(Campaign campaign, CurrencyType currency, String category, Long amount) {
